docs(table): extract anchorTableExample to its own file
